requirements.txt
matplotlib
numpy
pandas
seaborn
pytorch==1.1.0
IPython
sklearn
scikit-image==0.16.2